+++ MoEngagePluginCards/Helpers/MoEngagePluginCardsDelegateHandler.swift
@@ -1,0 +1,15 @@
+//
+//  MoEngagePluginCardsDelegate.swift
+//  MoEngagePluginCards
+//
+//  Created by Rakshitha on 16/08/23.
+//
+
+import Foundation
+import MoEngagePluginBase
+
+class MoEngagePluginCardsDelegateHandler: MoEngageCardsPluginDelegate {
+    override func initializeCardsPlugin() {
+       MoEngagePluginCardsBridge.sharedInstance
+    }
+}